fix: loading machine info no longer blocks reporting
