+++ src/main/java/io/snyk/agent/jvm/MachineInfo.java
@@ -1,0 +1,71 @@
+package io.snyk.agent.jvm;
+
+import java.io.*;
+import java.lang.management.ManagementFactory;
+import java.lang.management.RuntimeMXBean;
+import java.nio.file.Files;
+import java.util.Optional;
+
+public class MachineInfo {
+    public final String vmName;
+    public final String vmVendor;
+    public final String vmVersion;
+    public final String hostName = computeHostName();
+
+    {
+        final RuntimeMXBean runtime = ManagementFactory.getRuntimeMXBean();
+        vmName = runtime.getName();
+        vmVendor = runtime.getVmVendor();
+        vmVersion = runtime.getVmVersion();
+    }
+
+    private static String computeHostName() {
+        for (String env : new String[]{"SNYK_MACHINE_NAME", "HOSTNAME", "COMPUTERNAME"}) {
+            final String val = System.getenv(env);
+            if (null != val) {
+                final String trim = val.trim();
+                if (!trim.isEmpty()) {
+                    return trim;
+                }
+            }
+        }
+
+        try {
+            final Optional<String> line = Files.lines(new File("/etc/hostname").toPath()).findFirst();
+            if (line.isPresent()) {
+                final String trim = line.get().trim();
+                if (!trim.isEmpty()) {
+                    return trim;
+                }
+            }
+        } catch (IOException ignored) {
+            // no such file, couldn't read it, etc.
+            // e.g. docker, wrong OS (Windows)
+        }
+
+        try {
+            final String line = runHostname();
+            if (null != line) {
+                final String trim = line.trim();
+                if (!trim.isEmpty()) {
+                    return trim;
+                }
+            }
+        } catch (IOException ignored) {
+            // no such command, no permission to execute it, etc.
+        }
+
+        return "unknown";
+    }
+
+    private static String runHostname() throws IOException {
+        final Process proc = new ProcessBuilder("hostname").start();
+        proc.getOutputStream().close();
+        proc.getErrorStream().close();
+        try (final InputStream input = proc.getInputStream();
+             final InputStreamReader reader = new InputStreamReader(input);
+             final BufferedReader buffered = new BufferedReader(reader)) {
+            return buffered.readLine();
+        }
+    }
+}
